CRUX-1008: Fix infinite loading indicator bug
We were trying to turn off the loading indicator in the HTTP
Observable.do() method's onComplete() callback, but that
does not get called in the manner expected.  Moved the
calls into the onNext() and onError() callbacks instead.

Tried to use Observable.finally() to remove the duplication,
which worked in real life, but I couldn't figure out how to
make the specs detect that the proper method had been called.
So, I gave up on that after a reasonable effort.