CRUX-1309 & others: Store player state in Frames
The current time, clip duration, and in/out points
are now stored in the player state as Frame objects.
The player state now also includes a "changeDetectionEnabler"
property (which is set to the current time in milliseconds)
so that Angular components know that the state changed.

There is also a simple pipe called "timecode" that
converts a Frame object to a timecode, used for
display in player subcomponents.  (This could easily
be extended in the future to allow different time
formats in displays.)