CRUX-1615: Fix bug introduced during refactoring
I chose to have WzSubclipEditorComponent use
WzAdvancedPlayerComponent's "onUpdateSubclipData"
event emitter to determine when the player markers
changed, but that emitter doesn't fire when markers
are cleared, which meant that the Remove button never
showed up.

Added a new "markerChange" emitter which outputs
if and only if one marker (or both markers) change
from the previous player state.  Now WzSubclipEditorComponent
is properly kept up to date.