CRUX-2004: Move repeated component code into service
Specifically, this is code that converts a full asset
into just the parameters needed when saving it to a cart
or a quote.  Common code like this should be pushed
down to the service instead of repeated in many
components (and the component shouldn't care what
the backend wants in the cart or quote anyway).