CRUX-2660: Always clone UI config to protect it from changes
This is specifically to fix a problem where comment editing
was corrupting the comment form config by always setting the comment
field's value to the edited comment's previous value.  (Now that will
happen to a clone of the config, which should keep the original intact.)

But beyond that specific fix, this is a general solution to keep
all UI code from accidentally corrupting the UI config.